changed the appearnece of the edit_add products images

+++ products/views.py
@@ -15,21 +15,24 @@ def product_details(request,id):
         current_category = Category.objects.get(id = id)
         return render(request,'products/product_details.html',{'products':product,'category':category,'current':current_category})
 
-def add_product(request):
+def add_product(request,id):
     if 'admin_username' in request.session:
         if request.method == 'POST':
             name = request.POST.get('name')
-            id = request.POST.get('id')
-            category = Category.objects.get(id=id)
+            category = Category.objects.get(id = id)
             price = request.POST.get('price')
             stock = request.POST.get('stock')
-            images = request.FILES.getlist('images')
+            image1 = request.FILES.get('images1')
+            image2 = request.FILES.get('images2')
+            image3 = request.FILES.get('images3')
             discription = request.POST.get('discription')
             
             current = Products.objects.create(name=name,category=category,price = price, stocks=stock, discription = discription)
             
-            for image in images:
-                ProductImage.objects.create(product = current,image = image)
+            ProductImage.objects.create(product = current,image = image1)
+            ProductImage.objects.create(product = current,image = image2)
+            ProductImage.objects.create(product = current,image = image3)
+
 
             return redirect('product_details',id)
 
@@ -51,13 +54,15 @@ def edit_product(request,id):
             product.category = Category.objects.get(id = id)
             product.price = request.POST.get('price')
             product.stocks = request.POST.get('stock')
-            images = request.FILES.getlist('image')
+            if request.FILES.get('image'):
+                images = request.FILES.get('image')
+                img =ProductImage.objects.filter(product=product)
+                img.delete()
+                for image in images:
+                    ProductImage.objects.create(product = product,image = image)
             product.discription = request.POST.get('discription')
             product.save()
-            img =ProductImage.objects.filter(product=product)
-            img.delete()
-            for image in images:
-                ProductImage.objects.create(product = product,image = image)
+            
 
             return redirect('product_details',id)
